+++ src/mxcp/cli/init.py
@@ -10,7 +10,7 @@ import yaml
 
 from mxcp.cli.utils import configure_logging, get_env_profile, output_error
 from mxcp.config.analytics import track_command_with_timing
-from mxcp.config.duckdb_session import create_duckdb_session
+from mxcp.executor.session.duckdb import create_duckdb_session
 from mxcp.config.site_config import load_site_config
 from mxcp.config.user_config import load_user_config
