+++ src/mxcp/cli/validate.py
@@ -10,7 +10,7 @@ from mxcp.cli.utils import (
     output_result,
 )
 from mxcp.config.analytics import track_command_with_timing
-from mxcp.config.execution_engine import create_execution_engine
+from mxcp.executor.engine import create_execution_engine
 from mxcp.config.site_config import load_site_config
 from mxcp.config.user_config import load_user_config
 from mxcp.endpoints.validate import validate_all_endpoints, validate_endpoint
