+++ src/mxcp/evals/runner.py
@@ -3,12 +3,12 @@ import time
 from typing import Any, cast
 
 from mxcp.config._types import SiteConfig, UserConfig
-from mxcp.config.execution_engine import create_execution_engine
+from mxcp.executor.engine import create_execution_engine
 from mxcp.config.site_config import find_repo_root
 from mxcp.endpoints._types import EndpointDefinition
 from mxcp.endpoints.loader import EndpointLoader
 from mxcp.evals.loader import discover_eval_files, load_eval_suite
-from mxcp.evals.tool_executor import EndpointToolExecutor
+from mxcp.executor.runners.tool import EndpointToolExecutor
 from mxcp.sdk.auth import UserContext
 from mxcp.sdk.evals import (
     ClaudeConfig,
