+++ src/mxcp/executor/__init__.py
@@ -1,0 +1,1 @@
+"""Execution layer for MXCP."""
