+++ src/mxcp/executor/session/__init__.py
@@ -1,0 +1,1 @@
+"""Session management for executors."""
