+++ src/mxcp/executor/session/duckdb.py
